Made sure the relevant scripts don't output illegal utf-8 sequences
Two of the included scripts, bin/restic-run-backup.sh and
bin/maint/restic-run-backup--force.sh, or those that spawn
the 'restic backup' subcommand have been problematic in that
Restic occasionally generates illegal utf-8 sequences in its own status
line type of output when it truncates filenames to a certain lenght and
then issues a carriage return '\r' or a newline '\n' characted. Some
applications like Apache Airflow don't like this happening when
they're calling out to user supplied unix commands (for instance via
the BashOperator in Airflow) and logging the output in some fashion.
Without delving into Restic's status output code, the easy way to address
this issue is to filter Restic's output with a call to '| iconv -c -t UTF-8'
in the pipeline which strips out the offending byte sequences.