Accommodate Restic 0.10.0 changes in return code values
Swallows the `restic backup` return code 3 and returns 0 instead while warning that this signifies
that some files could not backed up. The 0.10.x series of Restic works differently compared to
previous 0.9.x series in that it returns a code of 3 in case of an incomplete backup. That means
some files could not be accessed due to files being open or otherwise locked which is very common on
Windows. Because I don't want to put every single problematic file in my user profile folder into
the exclusion list just to make sure the backup scripts return a 0 (i.e. no error) in order to not
alarm my task scheduler (Apache Airflow), I treat this condition as benign and put out a warning
instead. Your mileage may vary.